Fix the production build: move non-route exports out of a route module
`next build` failed on main:

    src/app/api/gateway/route.ts
    Type error: Route "src/app/api/gateway/route.ts" does not match the
    required types of a Next.js Route.
      "waitForResponsesEndpoint" is not a valid Route export field.

Next.js 16 allows a route module to export only handlers and its known
config fields, and the two chat routes imported these helpers from the
gateway route. Dev mode tolerates it, so the failure only appeared at build
time — the deployable artifact could not be produced at all.

Moved verbatim to `lib/responses-endpoint.ts`. Behaviour is unchanged: the
module still holds the cooldown state process-wide, so attempts stay
rate-limited and cannot become a gateway restart loop.
