+++ src/app/api/chat/route.ts
@@ -1,6 +1,6 @@
 import { runOpenResponsesText, guessMime } from "@/lib/openresponses";
 import { getGatewayUrl, getGatewayToken } from "@/lib/paths";
-import { waitForResponsesEndpoint, triggerResponsesEndpointSetup } from "@/app/api/gateway/route";
+import { triggerResponsesEndpointSetup, waitForResponsesEndpoint } from "@/lib/responses-endpoint";
 
 /**
  * Chat endpoint that sends a message to an OpenClaw agent and returns the response.
